feature: Add Weather module
This adds the `Weather` module which will be used to interface with the Dark Sky API. As of now, it only offers one method, `.temperature_of`. This method fetches the current temperature for the given zip code. If the zip code isn't found in our dataset (`lib/assets/zip-codes-to-lat-lng.yml`), an exception is raised.