feature: Set application time zone
For now, we're setting the application time zone to Pacific Time.

Currently, the only thing the time zone is used for is to define what today is so we can determine if a particular alert has been delivered today or not.

Setting the application time zone to Pacific Time should make the application work correctly for anyone in the U.S. To start, we're only supporting U.S. users anyways.

Pacific Time is basically the central time zone for the U.S. Every other time zone should be within 3 hours of Pacific Time. (Maybe up to 4 hours depending on Daylight Savings; I'm not sure.). This would mean that, at the earliest, the "day" would end at 8pm or 9pm for other time zones. By that time of day, temperatures are dropping so alerts shouldn't erroneously be delivered twice. At the latest, the "day" would end at 3am or 4am. Again, this shouldn't cause problems with sending alerts.

Later, we will likely change the application to be based on the specific time zone of each user. For now, I feel like the above implementation is a good starting point.