feature: Handle Dark Sky API Error Condition
According to the [Dark Sky API docs](https://darksky.net/dev/docs#flags), the Dark Sky response may include a `darksky-unavailable` flag. The presence of this property indicates that the Dark Sky data source supports the given location, but a temporary error (such as a radar station being down for maintenance) has made the data unavailable.

These changes check for this flag and raise an exception if it is found in the response.