tests: Move Webmock stub from support file to spec file
We ran into an issue where a request wasn't getting stubbed properly about half the time depending on what order you ran the specs in. When it wasn't stubbed properly, one of the specs in the `weather_spec.rb` file would fail because it attempted to make a real HTTP request.

I'm assuming it has something to do with how we were mocking the request in the support file. If you ran only the `weather_spec.rb` file, the spec would pass every time. If another spec file was loaded and executed before the `weather_spec.rb` file, the spec in the `weather_spec.rb` file would fail every time.

I didn't figure exactly what the problem was. I ultimately decided to move the Webmock stub from the support file into the `weather_spec.rb` file. This fixed the issue, and it also has the advantage of making it easier to understand the specs. I wish I would have figured out what was going on, but I like this solution well enough.